backend/requirements.txt
asgiref==3.2.7
autopep8==1.5.3
Django==3.0.6
djangorestframework==3.11.0
drf-writable-nested==0.6.0
paho-mqtt==1.5.0
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
